Create basic model-card-toolkit package
