+++ model_card_toolkit/documentation/examples/cats_vs_dogs_test.py
@@ -1,6 +1,7 @@
 """Tests for model_card_toolkit.documentation.examples.cats_vs_dogs."""
 
 from absl.testing import absltest
+
 from model_card_toolkit.documentation.examples import cats_vs_dogs
 
 SLICES = ['combined', 'cat', 'dog']
