+++ model_card_toolkit/model_card_test.py
@@ -16,13 +16,14 @@
 import json
 import os
 import pkgutil
+
 from absl.testing import absltest
+from google.protobuf import text_format
 import jsonschema
 
 from model_card_toolkit import model_card
 from model_card_toolkit.proto import model_card_pb2
 
-from google.protobuf import text_format
 
 _FULL_PROTO_FILE_NAME = "full.pbtxt"
 _FULL_PROTO = pkgutil.get_data(
