+++ model_card_toolkit/utils/graphics_utils.py
@@ -11,4 +11,26 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""A module of util functions for the Model Card Toolkit."""
+"""Utilities for generating plots and graphics."""
+
+import base64
+import io
+
+try:
+  from matplotlib.figure import Figure
+except:
+  Figure = None
+
+
+def figure_to_base64str(fig: Figure) -> str:
+  """Converts a Matplotlib figure to a base64 string encoding.
+
+  Args:
+    fig: A matplotlib Figure.
+
+  Returns:
+    A base64 encoding of the figure.
+  """
+  buf = io.BytesIO()
+  fig.savefig(buf, bbox_inches='tight', format='png')
+  return base64.b64encode(buf.getbuffer().tobytes()).decode('ascii')
