+++ model_card_toolkit/utils/testdata/tfxtest.py
@@ -16,16 +16,17 @@
 import os
 from typing import Any, Callable, List, Optional
 
+from absl import flags
 import apache_beam as beam
-from model_card_toolkit.utils.tfx_util import _TFX_METRICS_TYPE
-from model_card_toolkit.utils.tfx_util import _TFX_STATS_TYPE
+import ml_metadata as mlmd
+from ml_metadata.proto import metadata_store_pb2
+from tensorflow_metadata.proto.v0 import statistics_pb2
 import tensorflow_model_analysis as tfma
 from tensorflow_model_analysis.eval_saved_model.example_trainers import fixed_prediction_estimator
 from tfx_bsl.tfxio import raw_tf_record
 
-import ml_metadata as mlmd
-from ml_metadata.proto import metadata_store_pb2
-from tensorflow_metadata.proto.v0 import statistics_pb2
+from model_card_toolkit.utils.tfx_utils import _TFX_METRICS_TYPE
+from model_card_toolkit.utils.tfx_utils import _TFX_STATS_TYPE
 
 
 class TfxTest(tfma.eval_saved_model.testutil.TensorflowModelAnalysisTest):
@@ -174,3 +175,8 @@ class TfxTest(tfma.eval_saved_model.testutil.TensorflowModelAnalysisTest):
 
     if store:
       self._put_artifact(store, _TFX_STATS_TYPE, tfdv_path)
+
+if not __name__ == '__main__':
+  # Manually pass and parse flags to prevent UnparsedFlagAccessError when using
+  # pytest or unittest as a runner.
+  flags.FLAGS(['--test_tmpdir'])
